requirements.txt
Babel==2.8.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==2.2.16
django-allauth==0.43.0
django-appconf==1.0.4
django-bootstrap-form==3.4
django-classified==0.9
django-cors-headers==3.6.0
django-crispy-forms==1.9.2
django-filter==2.4.0
django-imagekit==4.0.2
django-rest-auth==0.9.5
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
pilkit==2.0
Pillow==8.0.1
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.4.1
Unidecode==1.1.1
urllib3==1.25.11
django-rest-swagger==2.2.0
